move to join folder
