+++ src/daft-distributed/src/pipeline_node/join/broadcast_join.rs
@@ -8,11 +8,10 @@ use daft_logical_plan::{stats::StatsState, JoinType};
 use daft_schema::schema::SchemaRef;
 use futures::{StreamExt, TryStreamExt};
 
-use super::{DistributedPipelineNode, SubmittableTaskStream};
 use crate::{
     pipeline_node::{
-        make_in_memory_scan_from_materialized_outputs, NodeID, NodeName, PipelineNodeConfig,
-        PipelineNodeContext,
+        make_in_memory_scan_from_materialized_outputs, DistributedPipelineNode, NodeID, NodeName,
+        PipelineNodeConfig, PipelineNodeContext, SubmittableTaskStream,
     },
     scheduling::{
         scheduler::{SchedulerHandle, SubmittableTask},
