+++ src/daft-distributed/src/pipeline_node/join/mod.rs
@@ -1,0 +1,6 @@
+mod broadcast_join;
+pub(crate) mod hash_join;
+pub(crate) mod translate_join;
+
+pub(crate) use broadcast_join::BroadcastJoinNode;
+pub(crate) use hash_join::HashJoinNode;
